Revert "feat: bridge banners (#365)"
This reverts commit 3a056ad50be83945f76e6732e120705507a7a669.
