+++ src/icons/index.tsx
@@ -19,7 +19,7 @@ import {
   ArrowLeft as ArrowLeftIcon,
   ArrowRight as ArrowRightIcon,
   ArrowUp as ArrowUpIcon,
-  ArrowUpRight as ArrowUpRightIcon,
+  ArrowUpRight as LinkIcon,
   BarChart2 as BarChart2Icon,
   ChevronDown as ChevronDownIcon,
   Clock as ClockIcon,
@@ -92,6 +92,7 @@ export const Clock = icon(ClockIcon)
 export const HelpCircle = icon(HelpCircleIcon)
 export const Identicon = icon(IdenticonIcon)
 export const Info = icon(InfoIcon)
+export const Link = icon(LinkIcon)
 export const AutoRouter = icon(RouterIcon)
 export const Settings = icon(SettingsIcon)
 export const Slash = icon(SlashIcon)
@@ -100,7 +101,6 @@ export const Wallet = icon(WalletIcon)
 export const X = icon(XIcon)
 export const XOctagon = icon(XOctagonIcon)
 export const Reverse = icon(ReverseIcon)
-export const ArrowUpRight = icon(ArrowUpRightIcon)
 
 export const Check = styled(icon(CheckIcon))`
   circle {
